context: Use uint32_t for context level. (#9502)
For this, it was checked if a negative level is ever used, and it is not,
also not in `popto`. The only special case is theory arith, which initializes
d_lastContextIntegerAttempted with -1, hence the special handling.

+++ src/prop/minisat/core/Solver.cc
@@ -394,7 +394,8 @@ CRef Solver::reason(Var x) {
   if (needProof() && explLevel < assertionLevel)
   {
     Trace("pf::sat") << "..user level is " << userContext()->getLevel() << "\n";
-    Assert(userContext()->getLevel() == (assertionLevel + 1));
+    Assert(userContext()->getLevel()
+           == static_cast<uint32_t>(assertionLevel + 1));
     d_proxy->notifyCurrPropagationInsertedAtLevel(explLevel);
   }
   // Construct the reason
